Flask_App/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.10.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
berserker-resolver==2.0.1
blinker==1.6.2
Bootstrap-Flask==2.2.0
bs4==0.0.1
build==0.10.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
config==0.5.1
crtsh==0.3.1
dnspython==2.3.0
# dnspython3==1.15.0
docopt==0.6.2
fastjsonschema==2.17.1
Flask==2.2.3
Flask-Cors==3.0.10
flask-paginate==2022.1.8
frozenlist==1.3.3
future==0.18.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_core==5.3.1
MarkupSafe==2.1.2
mongo==0.2.0
multidict==6.0.4
nbformat==5.9.0
nslookup==1.7.0
numpy==1.24.2
packaging==23.1
pandas==2.0.0
pathvalidate==2.5.2
pigar==2.0.8
pip-tools==6.13.0
pipreqs==0.4.13
platformdirs==3.6.0
pymongo==4.3.3
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-slugify==8.0.1
python-whois==0.8.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
# pywin32==306
requests==2.28.2
six==1.16.0
slugify==0.0.1
soupsieve==2.4.1
text-unidecode==1.3
traitlets==5.9.0
tzdata==2023.3
tzlocal==4.3
unicodecsv==0.14.1
Unidecode==1.3.6
urllib3==1.26.15
waybackpy==3.0.6
Werkzeug==2.2.3
WTForms==3.0.1
yarg==0.1.9
yarl==1.9.2
